interpolation and such

+++ engine/src/dsp.rs
@@ -2,5 +2,5 @@ pub trait Producer{
 
     fn amplitude(&mut self, sample: usize) -> f64;
 
-    fn sample_rate(&self) -> u32;
+    fn sample_rate(&self) -> usize;
 }
